refactor: refactored android modules structure
