+++ packages/react-native-audio-api/android/src/main/java/com/swmansion/audioapi/AudioAPIModule.kt
@@ -1,16 +1,20 @@
 package com.swmansion.audioapi
 
 import com.facebook.jni.HybridData
+import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactMethod
+import com.facebook.react.bridge.ReadableArray
+import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.common.annotations.FrameworkAPI
 import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.turbomodule.core.CallInvokerHolderImpl
+import com.swmansion.audioapi.system.MediaSessionManager
 
 @OptIn(FrameworkAPI::class)
 @ReactModule(name = AudioAPIModule.NAME)
 class AudioAPIModule(
-  reactContext: ReactApplicationContext,
+  val reactContext: ReactApplicationContext,
 ) : NativeAudioAPIModuleSpec(reactContext) {
   companion object {
     const val NAME = NativeAudioAPIModuleSpec.NAME
@@ -35,10 +39,48 @@ class AudioAPIModule(
     }
   }
 
-  @ReactMethod(isBlockingSynchronousMethod = true)
   override fun install(): Boolean {
+    MediaSessionManager.initialize(reactContext)
     injectJSIBindings()
 
     return true
+  }
+
+  override fun setLockScreenInfo(info: ReadableMap?) {
+    MediaSessionManager.setLockScreenInfo(info)
+  }
+
+  override fun resetLockScreenInfo() {
+    MediaSessionManager.resetLockScreenInfo()
+  }
+
+  override fun enableRemoteCommand(name: String?, enabled: Boolean) {
+    MediaSessionManager.enableRemoteCommand(name!!, enabled)
+  }
+
+  override fun setAudioSessionOptions(category: String?, mode: String?, options: ReadableArray?) {
+    // noting to do here
+  }
+
+  override fun getDevicePreferredSampleRate(): Double {
+    return MediaSessionManager.getDevicePreferredSampleRate()
+  }
+
+  override fun observeAudioInterruptions(enabled: Boolean) {
+    MediaSessionManager.observeAudioInterruptions(enabled)
+  }
+
+  override fun observeVolumeChanges(enabled: Boolean) {
+    MediaSessionManager.observeVolumeChanges(enabled)
+  }
+
+  override fun requestRecordingPermissions(promise: Promise?) {
+    val res = MediaSessionManager.requestRecordingPermissions(currentActivity)
+    promise!!.resolve(res)
+  }
+
+  override fun checkRecordingPermissions(promise: Promise?) {
+    val res = MediaSessionManager.checkRecordingPermissions()
+    promise!!.resolve(res)
   }
 }
