+++ packages/react-native-audio-api/android/src/main/java/com/swmansion/audioapi/AudioAPIPackage.kt
@@ -10,7 +10,6 @@ import com.facebook.react.module.model.ReactModuleInfoProvider
 @ReactModuleList(
   nativeModules = [
     AudioAPIModule::class,
-    AudioManagerModule::class,
   ],
 )
 class AudioAPIPackage : BaseReactPackage() {
@@ -20,7 +19,6 @@ class AudioAPIPackage : BaseReactPackage() {
   ): NativeModule? {
     when (name) {
       AudioAPIModule.NAME -> return AudioAPIModule(reactContext)
-      AudioManagerModule.NAME -> return AudioManagerModule(reactContext)
     }
     return null
   }
@@ -40,16 +38,6 @@ class AudioAPIPackage : BaseReactPackage() {
           isTurboModule = isTurboModule,
         )
 
-      moduleInfos[AudioManagerModule.NAME] =
-        ReactModuleInfo(
-          AudioManagerModule.NAME,
-          AudioManagerModule.NAME,
-          canOverrideExistingModule = true,
-          needsEagerInit = false,
-          hasConstants = true,
-          isCxxModule = false,
-          isTurboModule = false,
-        )
       moduleInfos
     }
 }
